+++ NEWS.md
@@ -1,8 +1,12 @@
+# gibble 0.2.0
+
+* Add gibble methods for silicate `PATH` and `PATH0`. 
+
 # gibble 0.1.2
 
 * Fix gibble for PATH objects from silicate (when they are points). 
 
-* SpatialPoints was has object column, needed for silicate. 
+* SpatialPoints has object column, needed for silicate. 
 
 # gibble 0.1.0